+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Added equals method to all data type classes.
 - Added toInteger method to IPAddress class.
+- Added fromInteger method to IPAddress class.
 
 ## [1.1.0] - 2017-10-01
 - Added EmailAddress class.